PHP 8.4 and 8.5 compatibility fixes

+++ generator/lib/builder/sql/DataSQLBuilder.php
@@ -202,7 +202,7 @@ abstract class DataSQLBuilder extends DataModelBuilder
      */
     protected function getDoubleSql($value)
     {
-        return (double) $value;
+        return (float) $value;
     }
 
     /**
